requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
absl-py==0.9.0            # via tensorboard, tensorflow
astor==0.8.1              # via tensorflow
bokeh==2.0.2              # via -r requirements.in
boto3==1.13.19            # via transformers
botocore==1.16.19         # via boto3, s3transfer
cachetools==4.1.0         # via google-auth
certifi==2020.4.5.1       # via requests
chardet==3.0.4            # via requests
click==7.1.2              # via nltk, sacremoses
docutils==0.15.2          # via botocore
faiss-gpu==1.6.3          # via -r requirements.in
filelock==3.0.12          # via transformers
future==0.18.2            # via pytorch-lightning, torch
gast==0.2.2               # via tensorflow
google-auth-oauthlib==0.4.1  # via tensorboard
google-auth==1.14.1       # via google-auth-oauthlib, tensorboard
google-pasta==0.2.0       # via tensorflow
grpcio==1.28.1            # via tensorboard, tensorflow
h5py==2.10.0              # via keras-applications
idna==2.9                 # via requests
jinja2==2.11.2            # via bokeh
jmespath==0.10.0          # via boto3, botocore
joblib==0.14.1            # via nltk, sacremoses, scikit-learn
keras-applications==1.0.8  # via tensorflow
keras-preprocessing==1.1.0  # via tensorflow
markdown==3.2.1           # via tensorboard
markupsafe==1.1.1         # via jinja2
nltk==3.5                 # via sentence-transformers
numpy==1.18.3             # via -r requirements.in, bokeh, faiss-gpu, h5py, keras-applications, keras-preprocessing, opt-einsum, pandas, pyarrow, pytorch-lightning, scikit-learn, scipy, sentence-transformers, tensorboard, tensorflow, tensorflow-hub, torch, transformers
oauthlib==3.1.0           # via requests-oauthlib
opt-einsum==3.2.1         # via tensorflow
packaging==20.3           # via bokeh
pandas==1.0.3             # via -r requirements.in
pillow==7.1.1             # via bokeh
protobuf==3.11.3          # via tensorboard, tensorflow, tensorflow-hub
pyarrow==0.17.0           # via -r requirements.in
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
pyparsing==2.4.7          # via packaging
python-dateutil==2.8.1    # via bokeh, botocore, pandas
pytorch-lightning==0.7.6  # via -r requirements.in
pytz==2019.3              # via pandas
pyyaml==5.3.1             # via bokeh, pytorch-lightning
regex==2020.5.14          # via nltk, sacremoses, transformers
requests-oauthlib==1.3.0  # via google-auth-oauthlib
requests==2.23.0          # via requests-oauthlib, tensorboard, transformers
rsa==4.0                  # via google-auth
s3transfer==0.3.3         # via boto3
sacremoses==0.0.43        # via transformers
scikit-learn==0.22.2.post1  # via -r requirements.in, sentence-transformers
scipy==1.4.1              # via scikit-learn, sentence-transformers, tensorflow
sentence-transformers==0.2.6.1  # via -r requirements.in
sentencepiece==0.1.91     # via transformers
six==1.14.0               # via absl-py, google-auth, google-pasta, grpcio, h5py, keras-preprocessing, packaging, protobuf, python-dateutil, sacremoses, tensorboard, tensorflow, tensorflow-hub
tensorboard==2.1.1        # via pytorch-lightning, tensorflow
tensorflow-estimator==2.1.0  # via tensorflow
tensorflow-hub==0.8.0     # via -r requirements.in
tensorflow-text==2.1.1    # via -r requirements.in
tensorflow==2.1.0         # via -r requirements.in, tensorflow-text
termcolor==1.1.0          # via tensorflow
tokenizers==0.5.2         # via transformers
torch==1.5.0              # via pytorch-lightning, sentence-transformers
tornado==6.0.4            # via bokeh
tqdm==4.45.0              # via -r requirements.in, nltk, pytorch-lightning, sacremoses, sentence-transformers, transformers
transformers==2.8.0       # via -r requirements.in, sentence-transformers
typing-extensions==3.7.4.2  # via bokeh
urllib3==1.25.9           # via botocore, requests
werkzeug==1.0.1           # via tensorboard
wheel==0.34.2             # via tensorboard, tensorflow
wrapt==1.12.1             # via tensorflow

# The following packages are considered to be unsafe in a requirements file:
# setuptools
